Add files via upload
A Logistic Regression project in Python to predict SUV purchases based on Age, Gender, and Estimated Salary. 
This repository contains code, dataset structure, and reports for building and evaluating the model. 
Please note: All materials are intended strictly for learning and demonstration purposes. 
To avoid plagiarism, kindly provide proper attribution and reference if you decide to use, modify, or extend this work in your own projects, 
presentations, or publications. Referencing the author and repository is encouraged when sharing.